+++ src/content/modules/index.ts
@@ -1,0 +1,3 @@
+import * as TUICPref from "./pref";
+
+export { TUICPref };
