+++ src/content/resources/buttonHTMLBase.html
@@ -1,0 +1,14 @@
+<div class="css-175oi2r TUICButtonUnderTweet TUICOriginalContent" style="display: inline-grid; justify-content: inherit; transform: rotate(0deg) scale(1) translate3d(0px, 0px, 0px); -moz-box-pack: inherit">
+    <div class="css-175oi2r r-18u37iz r-1h0z5md">
+        <div TUICButton="${id}" role="button" tabindex="${tabindex}" class="css-175oi2r r-1777fci r-bt1l66 r-1ny4l3l r-bztko3 r-lrvibr ${clazzBtn}">
+            <div dir="ltr" class="css-901oao r-1awozwy r-6koalj r-37j5jr r-a023e6 r-16dba41 r-1h0z5md r-bcqeeo r-o7ynqc r-clp7b1 r-3s2u2q r-qvutc0 ${clazzLtr} TUIC_ButtonHover2">
+                <div class="css-175oi2r r-xoduu5 TUIC_ButtonHover">
+                    <div class="css-175oi2r r-1niwhzg r-sdzlij r-1p0dtai r-xoduu5 r-1d2f490 r-xf4iuw r-1ny4l3l r-u8s1d r-zchlnj r-ipm5af r-o7ynqc r-6416eg"></div>
+                    <svg viewBox="0 0 24 24" aria-hidden="true" class="r-4qtqp9 r-yyyyoo r-1q142lx r-dnmrzs r-bnwqim r-1plcrui r-lrvibr ${clazzSVG}">
+                        <g>${svg}</g>
+                    </svg>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
