+++ src/content/resources/index.ts
@@ -1,0 +1,3 @@
+import buttonHTMLBase from "./buttonHTMLBase.html?raw";
+
+export { buttonHTMLBase };
