+++ src/shared/options/components/CheckBox.vue
@@ -1,6 +1,6 @@
 <template>
     <div class="TUICCheckBoxParent">
-        <input type="checkbox" :id="value.replace(/\./g, '-_-')" :checked="TUICPref.get(value)" @change="changePref(value, $event)" />
+        <input type="checkbox" :id="value.replace(/\./g, '-_-')" :checked="TUICPref.getPref(value)" @change="changePref(value, $event)" />
         <div>
             <label class="TUIC_setting_text" :for="value.replace(/\./g, '-_-')">{{ TUICI18N.get(name) }}</label>
         </div>
@@ -10,8 +10,9 @@
 <script lang="ts">
 import { defineComponent } from "vue";
 import { TUICI18N } from "@content/i18n";
-import { TUICPref, TUICLibrary } from "@content/library";
+import { TUICLibrary } from "@content/library";
 import { TUICObserver } from "@content/observer";
+import { TUICPref } from "@content/modules";
 
 export default defineComponent({
     props: {
@@ -26,7 +27,7 @@ export default defineComponent({
     },
     setup() {
         const changePref = (path, event) => {
-            TUICPref.set(path, event.target.checked);
+            TUICPref.setPref(path, event.target.checked);
             TUICPref.save();
             TUICLibrary.getClasses.update();
             TUICObserver.titleObserverFunction();
