+++ src/vite.d.ts
@@ -2,3 +2,8 @@ declare module "*?url" {
     let str: string;
     export = str;
 }
+
+declare module "*?raw" {
+    let str: string;
+    export = str;
+}
